implementacion de findAll, findById y create de employee

+++ src/main/java/com/humanResources/humanResourcesAPI/vo/PositionVo.java
@@ -3,13 +3,14 @@ package com.humanResources.humanResourcesAPI.vo;
 import com.humanResources.humanResourcesAPI.model.entity.Department;
 import com.humanResources.humanResourcesAPI.model.entity.Employee;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 public record PositionVo(
     Long id,
     String name,
     String descr,
-    Double baseSalary,
+    BigDecimal baseSalary,
     Department department,
     Set<Employee>employees
 ) {
